# docs/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "WeMail 菜单化后端 API",
    "version": "2.0.0",
    "description": "WeMail 后端 API 已按管理后台左侧菜单分组。旧 /auth、/admin、/api/mailboxes、/api/messages、/api/outbound、/api/keys、/api/telegram 路径不再保留。"
  },
  "servers": [
    {
      "url": "/",
      "description": "当前 Worker 服务根路径"
    }
  ],
...
  "paths": {
...
    "/api/api-keys": {
      "get": {
        "tags": [
          "API 密钥"
        ],
        "summary": "获取 API Key 列表",
        "operationId": "listApiKeys",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "API Key 列表"
          }
        }
      },
      "post": {
        "tags": [
          "API 密钥"
        ],
        "summary": "创建 API Key",
        "operationId": "createApiKey",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "创建成功"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/api-keys/{id}": {
      "delete": {
        "tags": [
          "API 密钥"
        ],
        "summary": "吊销 API Key",
        "operationId": "revokeApiKey",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPath"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          }
        }
      }
    },
...
    "/api/announcements": {
      "get": {
        "tags": [
          "公告"
        ],
        "summary": "获取公告列表",
        "operationId": "listAnnouncements",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "公告列表"
          }
        }
      },
      "post": {
        "tags": [
          "公告"
        ],
        "summary": "发布公告",
        "operationId": "createAnnouncement",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "发布成功"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/accounts": {
      "get": {
        "tags": [
          "账号"
        ],
        "summary": "获取当前用户账号列表",
        "operationId": "listAccounts",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "账号列表"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "tags": [
          "账号"
        ],
        "summary": "创建账号",
        "operationId": "createAccount",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/AccountCreateRequest"
        },
        "responses": {
          "201": {
            "description": "创建成功"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/accounts/{id}": {
      "delete": {
        "tags": [
          "账号"
        ],
        "summary": "删除账号",
        "operationId": "deleteAccount",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPath"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Ok"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/accounts/settings": {
      "get": {
        "tags": [
          "账号"
        ],
        "summary": "获取账号策略",
        "operationId": "getAccountSettings",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "账号策略"
          }
        }
      },
      "put": {
        "tags": [
          "账号"
        ],
        "summary": "更新账号策略",
        "operationId": "updateAccountSettings",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "账号策略"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/mail/messages": {
      "get": {
        "tags": [
          "邮件"
        ],
        "summary": "获取账号消息列表",
        "operationId": "listMailMessages",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "accountId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "消息列表"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/mail/messages/{id}": {
      "get": {
        "tags": [
          "邮件"
        ],
        "summary": "获取消息详情",
        "operationId": "getMailMessage",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "消息详情"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/mail/messages/{messageId}/attachments/{attachmentId}": {
      "get": {
        "tags": [
          "邮件"
        ],
        "summary": "获取消息附件",
        "operationId": "getMailAttachment",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "attachmentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "附件元数据或文件流"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/mail/outbound": {
      "get": {
        "tags": [
          "邮件"
        ],
        "summary": "获取发件记录",
        "operationId": "listMailOutbound",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "accountId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "发件记录"
          }
        }
      }
    },
    "/api/mail/send": {
      "post": {
        "tags": [
          "邮件"
        ],
        "summary": "发送邮件",
        "operationId": "sendMail",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "发送成功并返回配额"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
    "/api/mail/settings": {
      "get": {
        "tags": [
          "邮件"
        ],
        "summary": "获取邮件设置",
        "operationId": "getMailSettings",
        "security": [
          {
            "cookieAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "邮件设置"
          }
        }
      },
      "put": {
        "tags": [
          "邮件"
        ],
        "summary": "更新邮件设置",
        "operationId": "updateMailSettings",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "邮件设置"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "wemail_session"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "parameters": {
      "IdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "requestBodies": {
...
      "AccountCreateRequest": {
        "required": false,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "label": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "responses": {
      "Ok": {
        "description": "操作成功",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "ok": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      },
      "Error": {
        "description": "请求失败",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    }
  }
}